+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # MultiQC Version History
 
 #### v0.9dev
+Module updates:
+* [**Prokka**](http://www.vicbioinformatics.com/software.prokka.shtml) - new module!
+  * Prokka is a software tool for the rapid annotation of prokaryotic genomes.
 * **Cutadapt**
   * Fixed bug in General Stats table number for old versions of cutadapt (pre v1.7)
   * Added support for _really_ old cutadapt logs (eg. v.1.2)